uint to uint256

+++ contracts/swaps/HelixPair.sol
@@ -173,7 +173,7 @@ contract HelixPair is HelixLP, ReentrancyGuard {
     }
 
     // this low-level function should be called from a contract which performs important safety checks
-    function swap(uint amount0Out, uint amount1Out, address to) external nonReentrant {
+    function swap(uint256 amount0Out, uint256 amount1Out, address to) external nonReentrant {
         require(amount0Out > 0 || amount1Out > 0, "Helix INSUFFICIENT_OUTPUT_AMOUNT");
         (uint112 _reserve0, uint112 _reserve1,) = getReserves(); // gas savings
         require(amount0Out < _reserve0 && amount1Out < _reserve1, "Helix INSUFFICIENT_LIQUIDITY");
